refactor(note_graph): code separation
