+++ src/features/note_graph/mod.rs
@@ -1,5 +1,6 @@
 pub mod note_graph;
 pub mod note_graph_ui;
+mod note_graph_mock;
 
 pub use note_graph::*;
 pub use note_graph_ui::*;
